chore(deps): remove unnecessary cmake tl-expected shellcheck cmake

+++ include/leetcode.h
@@ -235,7 +235,7 @@ TreeNode *stringToTreeNode(string input) {
 string stringToString(string input) {
   assert(input.length() >= 2);
   string result;
-  for (int i = 1; i < input.length() - 1; i++) {
+  for (size_t i = 1; i < input.length() - 1; i++) {
     char currentChar = input[i];
     if (input[i] == '\\') {
       char nextChar = input[i + 1];
